GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Update frontend dockerfile (refs/gitbutler/Update-frontend-dockerfile)
   - .github/workflows/pulls.yml
   - package.json
   - pnpm-lock.yaml
   - babel.config.js
   - jest.config.js
   - jest.config.ts

Your previous branch was: refs/heads/main

The sha for that commit was: 79079a5ec1d333785787725e4acc15513215b415

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch